Migrate A-10C to new module type
