+++ Scripts/DCS-BIOS/doc/Addresses.h
@@ -3962,7 +3962,7 @@
 #define A_10C_VHFAM_FREQ2_ROT 0x12B8, 0xFF00, 8
 #define A_10C_VHFAM_FREQ3 0x118E, 0x0F00, 8
 #define A_10C_VHFAM_FREQ3_ROT 0x12C2, 0xFF00, 8
-#define A_10C_VHFAM_FREQ4 0x118E, 0x7000, 12
+#define A_10C_VHFAM_FREQ4 0x118E, 0x3000, 12
 #define A_10C_VHFAM_FREQ4_A 0x1192
 #define A_10C_VHFAM_FREQ4_ROT 0x12CA, 0xFF00, 8
 #define A_10C_VHFAM_FREQEMER 0x1186, 0x0C00, 10
@@ -3981,7 +3981,7 @@
 #define A_10C_VHFFM_FREQ2_ROT 0x1306, 0xFF00, 8
 #define A_10C_VHFFM_FREQ3 0x119C, 0x00F0, 4
 #define A_10C_VHFFM_FREQ3_ROT 0x1308, 0xFF00, 8
-#define A_10C_VHFFM_FREQ4 0x119C, 0x0700, 8
+#define A_10C_VHFFM_FREQ4 0x119C, 0x0300, 8
 #define A_10C_VHFFM_FREQ4_A 0x119E
 #define A_10C_VHFFM_FREQ4_ROT 0x1312, 0xFF00, 8
 #define A_10C_VHFFM_FREQEMER 0x1194, 0x0180, 7
