+++ Scripts/DCS-BIOS/src/json/A-10C_CDU.json
@@ -3333,6 +3333,17 @@
             "y": 1
         }
     ],
+    "Cursor": [
+        {
+            "alignment": "LFT",
+            "pages": [],
+            "id": "Cursor",
+            "index": 0,
+            "statictext": false,
+            "x": 2,
+            "y": 10
+        }
+    ],
     "DATA": [
         {
             "alignment": "LFT",
@@ -12542,6 +12553,28 @@
             "y": 4
         }
     ],
+    "PAGE": [
+        {
+            "alignment": "LFT",
+            "pages": [],
+            "id": "PAGE",
+            "index": 0,
+            "statictext": false,
+            "x": 19,
+            "y": 10
+        }
+    ],
+    "PAGE_NUM": [
+        {
+            "alignment": "LFT",
+            "pages": [],
+            "id": "PAGE_NUM",
+            "index": 0,
+            "statictext": false,
+            "x": 20,
+            "y": 10
+        }
+    ],
     "PAGE_NUMBER": [
         {
             "alignment": "LFT",
@@ -14015,6 +14048,28 @@
             "statictext": true,
             "x": 1,
             "y": 5
+        }
+    ],
+    "ScratchBorders": [
+        {
+            "alignment": "LFT",
+            "pages": [],
+            "id": "ScratchBorders",
+            "index": 0,
+            "statictext": false,
+            "x": 1,
+            "y": 10
+        }
+    ],
+    "ScratchString": [
+        {
+            "alignment": "LFT",
+            "pages": [],
+            "id": "ScratchString",
+            "index": 0,
+            "statictext": false,
+            "x": 2,
+            "y": 10
         }
     ],
     "SCS": [
